delete windows backend
turns out when using the UWP platform you cannot call functions such as CreateWindowEx. which is a problem because the emulator window needs to be a native type for the Core C++ emulator to be properly cross platform. so im going to have to rewrite the UI from scratch in normal win32 api style